Move RepoActionsExt out of gitbutler-repository
This is a higher level trait with lots of dependencies
