+++ crates/gitbutler-branch-actions/src/branch_manager/branch_removal.rs
@@ -7,7 +7,8 @@ use gitbutler_commit::commit_headers::CommitHeadersV2;
 use gitbutler_oplog::SnapshotExt;
 use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_reference::{normalize_branch_name, ReferenceName, Refname};
-use gitbutler_repo::{RepoActionsExt, RepositoryExt};
+use gitbutler_repo::RepositoryExt;
+use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::{Stack, StackId};
 use tracing::instrument;
