+++ crates/gitbutler-branch-actions/src/virtual.rs
@@ -25,8 +25,9 @@ use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_reference::{normalize_branch_name, Refname, RemoteRefname};
 use gitbutler_repo::{
     rebase::{cherry_rebase, cherry_rebase_group},
-    LogUntil, RepoActionsExt, RepositoryExt,
+    LogUntil, RepositoryExt,
 };
+use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::{
     reconcile_claims, BranchOwnershipClaims, Stack, StackId, Target, VirtualBranchesHandle,
 };
