+++ crates/gitbutler-repo-actions/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "gitbutler-repo-actions"
+version = "0.0.0"
+edition = "2021"
+authors = ["GitButler <gitbutler@gitbutler.com>"]
+publish = false
+
+[dependencies]
+git2.workspace = true
+serde = { workspace = true, features = ["std"] }
+tokio = { workspace = true, features = [
+    "rt-multi-thread",
+    "rt",
+    "macros",
+    "sync",
+] }
+log = "^0.4"
+anyhow = "1.0.86"
+gitbutler-command-context.workspace = true
+tracing.workspace = true
+gitbutler-stack.workspace = true
+gitbutler-id.workspace = true
+gitbutler-git.workspace = true
+gitbutler-project.workspace = true
+gitbutler-error.workspace = true
+gitbutler-commit.workspace = true
+gitbutler-reference.workspace = true
+gitbutler-repo.workspace = true
+gitbutler-time.workspace = true
