+++ crates/gitbutler-repo-actions/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod askpass;
+
+mod repository;
+pub use repository::RepoActionsExt;
