+++ crates/gitbutler-repo/src/lib.rs
@@ -1,20 +1,15 @@
 pub mod rebase;
 
-mod repository;
-pub use repository::{LogUntil, RepoActionsExt};
-
 mod commands;
 pub use commands::{FileInfo, RepoCommands};
 
 mod repository_ext;
-pub use repository_ext::{GixRepositoryExt, RepositoryExt};
+pub use repository_ext::{GixRepositoryExt, LogUntil, RepositoryExt};
 
 pub mod credentials;
 
 mod config;
 
 pub use config::Config;
-
-pub mod askpass;
 
 pub mod temporary_workdir;
