+++ crates/gitbutler-repo/src/repository_ext.rs
@@ -4,7 +4,7 @@ use std::os::unix::fs::PermissionsExt;
 use std::os::windows::process::CommandExt;
 use std::{io::Write, path::Path, process::Stdio, str};
 
-use crate::{Config, LogUntil};
+use crate::Config;
 use anyhow::{anyhow, bail, Context, Result};
 use bstr::BString;
 use git2::{BlameOptions, StatusOptions, Tree};
@@ -703,4 +703,20 @@ impl GixRepositoryExt for gix::Repository {
         self.object_cache_size_if_unset(bytes);
         Ok(self)
     }
+}
+
+type OidFilter = dyn Fn(&git2::Commit) -> Result<bool>;
+
+/// Generally, all traversals will use no particular ordering, it's implementation defined in `git2`.
+pub enum LogUntil {
+    /// Traverse until one sees (or gets commits older than) the given commit.
+    /// Do not return that commit or anything older than that.
+    Commit(git2::Oid),
+    /// Traverse the given `n` commits.
+    Take(usize),
+    /// Traverse all commits until the given condition returns `false` for a commit.
+    /// Note that this commit-id will also be returned.
+    When(Box<OidFilter>),
+    /// Traverse the whole graph until it is exhausted.
+    End,
 }
