+++ crates/gitbutler-stack-api/tests/mod.rs
@@ -3,7 +3,8 @@ use gitbutler_command_context::CommandContext;
 use gitbutler_commit::commit_ext::CommitExt;
 use gitbutler_patch_reference::{CommitOrChangeId, PatchReference};
 use gitbutler_reference::RemoteRefname;
-use gitbutler_repo::{LogUntil, RepoActionsExt, RepositoryExt as _};
+use gitbutler_repo::{LogUntil, RepositoryExt as _};
+use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::VirtualBranchesHandle;
 use gitbutler_stack_api::{PatchReferenceUpdate, StackExt, TargetUpdate};
 use itertools::Itertools;
